+++ api-runtime/grpc-runtime/service/vpc.go
@@ -11,7 +11,7 @@ package service
 import (
 	"context"
 	"strings"
-
+	cm "github.com/cloud-barista/cb-spider/api-runtime/common-runtime"
 	gc "github.com/cloud-barista/cb-spider/api-runtime/grpc-runtime/common"
 	"github.com/cloud-barista/cb-spider/api-runtime/grpc-runtime/logger"
 	pb "github.com/cloud-barista/cb-spider/api-runtime/grpc-runtime/stub/cbspider"
@@ -33,12 +33,12 @@ func (s *CCMService) CreateVPC(ctx context.Context, req *pb.VPCCreateRequest) (*
 	logger.Debug("calling CCMService.CreateVPC()")
 
 	// check the input Name to include the SUBNET: Prefix
-	if strings.HasPrefix(req.Item.Name, rsSubnetPrefix) {
-		return nil, gc.NewGrpcStatusErr(rsSubnetPrefix+" cannot be used for VPC name prefix!!", "", "CCMService.CreateVPC()")
+	if strings.HasPrefix(req.Item.Name, cm.SUBNET_PREFIX) {
+		return nil, gc.NewGrpcStatusErr(cm.SUBNET_PREFIX+" cannot be used for VPC name prefix!!", "", "CCMService.CreateVPC()")
 	}
 	// check the input Name to include the SecurityGroup Delimiter
-	if strings.HasPrefix(req.Item.Name, sgDELIMITER) {
-		return nil, gc.NewGrpcStatusErr(sgDELIMITER+" cannot be used in VPC name!!", "", "CCMService.CreateVPC()")
+	if strings.HasPrefix(req.Item.Name, cm.SG_DELIMITER) {
+		return nil, gc.NewGrpcStatusErr(cm.SG_DELIMITER+" cannot be used in VPC name!!", "", "CCMService.CreateVPC()")
 	}
 
 	// Grpc RegInfo => Driver ReqInfo
@@ -182,7 +182,7 @@ func (s *CCMService) AddSubnet(ctx context.Context, req *pb.SubnetAddRequest) (*
 	reqSubnetInfo := cres.SubnetInfo{IId: cres.IID{req.Item.Name, ""}, IPv4_CIDR: req.Item.Ipv4Cidr}
 
 	// Call common-runtime API
-	result, err := cmrt.AddSubnet(req.ConnectionName, rsSubnetPrefix+req.VpcName, req.VpcName, reqSubnetInfo)
+	result, err := cmrt.AddSubnet(req.ConnectionName, cm.SUBNET_PREFIX+req.VpcName, req.VpcName, reqSubnetInfo)
 	if err != nil {
 		return nil, gc.ConvGrpcStatusErr(err, "", "CCMService.AddSubnet()")
 	}
@@ -205,7 +205,7 @@ func (s *CCMService) RemoveSubnet(ctx context.Context, req *pb.SubnetQryRequest)
 	logger.Debug("calling CCMService.RemoveSubnet()")
 
 	// Call common-runtime API
-	result, _, err := cmrt.DeleteResource(req.ConnectionName, rsSubnetPrefix+req.VpcName, req.SubnetName, req.Force)
+	result, _, err := cmrt.DeleteResource(req.ConnectionName, cm.SUBNET_PREFIX+req.VpcName, req.SubnetName, req.Force)
 	if err != nil {
 		return nil, gc.ConvGrpcStatusErr(err, "", "CCMService.RemoveSubnet()")
 	}
@@ -221,7 +221,7 @@ func (s *CCMService) RemoveCSPSubnet(ctx context.Context, req *pb.CSPSubnetQryRe
 	logger.Debug("calling CCMService.RemoveCSPSubnet()")
 
 	// Call common-runtime API
-	result, _, err := cmrt.DeleteCSPResource(req.ConnectionName, rsSubnetPrefix+req.VpcName, req.Id)
+	result, _, err := cmrt.DeleteCSPResource(req.ConnectionName, cm.SUBNET_PREFIX+req.VpcName, req.Id)
 	if err != nil {
 		return nil, gc.ConvGrpcStatusErr(err, "", "CCMService.RemoveCSPSubnet()")
 	}